moved db controller functions to specific controllers
