+++ api/routes/time_routes.py
@@ -1,17 +1,17 @@
 from bottle import Bottle, request, abort
-from controllers.db_controller import db_controller
+from controllers.height_controller import height_controller
 
 time_server = Bottle()
 
 
 @time_server.route("/today")
 def get_todays_total_time():
-    return db_controller.get_todays_total()
+    return height_controller.get_todays_total()
 
 
 @time_server.route("/yesterday")
 def get_todays_total_time():
-    return db_controller.get_yesterdays_total()
+    return height_controller.get_yesterdays_total()
 
 @time_server.route("/year")
 def get_todays_total_time():
@@ -20,10 +20,10 @@ def get_todays_total_time():
     if year is None:
         abort(400)
     
-    return db_controller.get_daily_totals_of_year(year)
+    return height_controller.get_daily_totals_of_year(year)
 
 @time_server.route("/day")
 def get_todays_total_time():
     day = request.headers.get("day")
 
-    return db_controller.get_totals_of_day(day)
+    return height_controller.get_totals_of_day(day)
